melhorias na saida do avaliador

+++ src/avaliador.h
@@ -1,13 +1,14 @@
 #ifndef __AVALIADOR__H__
 #define __AVALIADOR__H__ value
 
+#include <cstdint>
 
 class Search;
 class labirinto;
 
 class avaliador {
     private:
-        double tempoDecorrido(Search *busca);
+         uint64_t tempoDecorrido(Search *busca);
     public:
         avaliador();
         ~avaliador();
